requirements.txt
MouseInfo==0.1.3
numpy==2.3.1
pillow==11.2.1
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
pytweening==1.2.0
pywin32==310
